fix: reduce GPU memory in k_summary computation
Change native_recent_k.float().mean() to compute in native dtype first,
then convert result back. Avoids allocating full float32 copy of the
recent K tensor (~200MB for 32760 tokens × 12 heads × 128 dim).

+++ src/lifecycle_kv/runtime.py
@@ -361,7 +361,8 @@ class LifeCacheRuntime:
             k=native_recent_k,
             v=native_recent_v,
             token_indices=token_indices,
-            k_summary=native_recent_k.float().mean(dim=0),
+            # Compute summary in native dtype to avoid OOM from float32 conversion
+            k_summary=native_recent_k.to(torch.float32).mean(dim=0).to(native_recent_k.dtype),
             region=CacheRegion.RECENT,
         )
